FIX: default GitBlit config is more Apache Proxy-friendly.
Apache reverse proxy does not tolerate well the '/' encoded
in the URL string as %2F: it does the translation before the
proxy rule is applied causing confusion to GitBlit parameters
mounted as URI path.

Disabling the mounting of parameters in the URI path makes
GitBlit plugin more "Apache-proxy" friendly and avoiding
lots of headaches.

Change-Id: I591be9e375858012916c0df4f55d3ad0d022b148
Signed-off-by: Luca Milanesio <luca.milanesio@gmail.com>