web.otherUrls auto-config based on Gerrit SSH/HTTPD URLs.
Allow GitBlit to automatically inherit the Gerrit SSH
and HTTPD URLs configuration into its web.otherUrls
settings.

This allows to have in GitBlit the "repository URL"
fields correctly populated and customised on a per
user and per repo basis.

Additionally the URLs, Git, SourceTree and Tower
client buttons will be automatically populated and
working out-of-the-box.

Change-Id: I447768919039e1b3737c7137ce8396ff86b91a1d