Fixed 2.8-SNAPSHOT issue with web.otherUrls
Fixed the issue of web.otherUrls, where now this
property is populated from gerrit.config file
and from the properties.gitblit file. In order
to show username with repository URLs, you need
to add {1} to denote the username and {0} to
denote the project name in properties file.

As end user for gitblit plugin, you can add
gerrit mirror URLs to the list of URLs of repo
in the gitblit UI,e.g.
web.otherUrls = ssh://{1}@gerritmirror1:29418/{0}
ssh://{1}@gerritmirror2:29418/{0}

Before this fix, the property was populated only
from gerrit.config file preventing adding mirror
to available URLs in gitblit plugin UI.

If properties file does not contain web.otherUrls
the existing value of the variable is not added to
web.otherUrls, to avoid adding NULL to it.

Change-Id: I0130022817d941469086a49e2148ff9ce70a90f1