Get rid of hacked Wicket/Rome versions
Let's use standard distributions instead.

Rome is easy, it's used only in GitBlit's SyndicationServlet: we can
just set and reset the class loader appropriately.

Wicket is more complicated, but from looking at the Wicket code and the
original hack at
https://github.com/lucamilanesio/wicket/commit/4a3f2c1dccf064008f223bb554c90e12c1c3b55d
it appears to me that this hack may not have been necessary in the first
place and just overriding getClassLoader() in the GerritWicketFilter
might be good enough.