Exclude groovy
GitBlit uses groovy only for its pre- and post-push hooks. Since we're
running it as a viewer, this is never used. We're not even getting
anyhere near anything that would want to instantiate a class that used
anything from groovy, so removing it is safe and saves another 6MB!

Also remove the GitBlit client and GitBlit GO from the jar.

Final jar size is now down to about 17MB. Still large, but at least in
the same ballpark as the old official Gerrit-GitBlit plugin.