Fix display of image files
Turns out that GitBlit needs some keys in its settings to know which
kinds of files are images or binary. Added them to the built-in
configuration, but made them overridable by user-defined settings.

Somehow I thought GitBlit used a Mime detector to figure this out on
its own, but apparently not.

Related: https://dev.gitblit.com/tickets/gitblit.git/133 and
https://code.google.com/p/gitblit/issues/detail?id=457