Login link instead of form
The GitBlit login form works only if Gerrit is configured to use a
username-password authentication scheme (basically LDAP). Thanks to
Luca Milanesio for pointing this out.

Therefore we only show a login _link_ instead of the form if Gerrit is
configured to use any other authentication scheme. The link goes to
Gerrit's /login page, which then will do the right thing.

We inspect Gerrit's config, and if it's not LDAP, we construct the URL
to Gerrit's login page, which we then transport to the RootPage via the
GitBlit settings.