Fix footer link on GitBlit pages going back to Gerrit.
If Gerrit was run not on / but on a subdirectory such as /review/
(for instance in a reverse-proxy setup), the footer link would still
go to /.

Fixed by setting that link dynamically to Gerrit's canonicalWebUrl,
which we include in GitBlit's settings so that it is available in
BasePage.

Cherry-picked from commit 331ff6b.