Issue #7: Fix gravatar failure with Github OAuth
Gitblit assumes that a UserModel:

- always has a username (dealt with in commit 524e7da)
- has either an e-mail address or a display name

The plugin never set the e-mail address, and through a logic error
also might not set the display name if the Gerrit account had none.

The fix ensures that the e-mail address and display name of the Gerrit
account are propagated to Gitblit. If the Gerrit account has no display
name, we use the username (guaranteed to be set since commit 524e7da)
instead.