Fix documentation
Documentation/index.md had a backslash-escaped \${file} in the sample
Gerrit configuration. That got replaced by the maven build, but
remained visible in the source, which the main README linked to.

This made some people insert that backslash into their gerrit.config,
which apparently made Gerrit fail to start. See PR #1.

Solved by using a custom delimiter for the maven resources plugin.

Additionally, add a build step that puts the processed documentation
page not only into the jar, but also in the project itself. Then link
to that from README.md