Respect the Gerrit preference for the number of context lines
GitBlit unconditionally uses a diff context of 3 lines, which may be a
tad small. Change this so that the plugin respects the value the user
has set in his or her Gerrit preferences.

Introduce a new GitBlit configuration property to limit the number of
context lines in a commit diff. Otherwise GitBlit might report too many
files as "Diff too large" in a diff over all files in a commit.

Update documentation and increase version number.