Issue #22: gitblit's bugtraq fails on empty repository
If a repository is completely empty (no initial commit), but a change
was pushed to Gerrit and one then tries to view that change in Gitblit,
Gitblit fails with an internal server error caused by an NPE in
Syntevo's BugtraqConfig, which tries to read some config file from the
repository.

Fix by handling the case of a repository not having a HEAD yet.