Issue #21: ensure that inner classes don't reference JGit objects
Gitblit uses a few inner anonymous classes in some pages. Ensure that
those do not hold references to JGit RevCommits to avoid that they get
serialized when the page is.

The usage of serialized inner anonymous classes is an anti-pattern in
Wicket programming and is discouraged in the Java serialization spec,
but cleaning this up properly would amount to a larger refactoring than
can be done in this plugin (and also larger than I'm willing to do).